Describe the new JIRA filters and custom fields for reviewing issues.
To encourage members of the community to help whittle down the open JIRA list for Grails, there are now a couple of new custom fields and a corresponding filter that make it easy to track what issues need reviewing and which ones the community feels can be closed. These features are now documented in the contributing section of the user guide.