Fix broken link to ConfigSlurper documentation from configuration guide
The link to ConfigSlurper is broken. The reason is because the author
used the markup:

   "ConfigSlurper":http://groovy.codehaus.org/ConfigSlurper.

Mixing the URL with punctuation (the trailing '.',  a valid URL
character) results in the '.' being included in the generated URL -
"http://groovy.codehaus.org/ConfigSlurper."

I switched up the markup used to be an alternative link markup:

   [ConfigSlurper|http://groovy.codehaus.org/ConfigSlurper].

The result is the output the author intended.