Fix command object documentation
The documentation says that if the command object is defined in the same file as the controller it does not need to be marked @Validateable. This is not true. The command object must be used as a parameter of the action that wants to use it to not need it to be marked as @Validateable. The error that one gets when attempting to use a command object that is defined in the same class as the controller, but not used as a parameter to the action being used is:
org.codehaus.groovy.grails.web.taglib.exceptions.GrailsTagException: Tag [validate] is missing required attribute [form]

If this is a bug, let me know and I will remove this pull request.