Clarify usage of the 'mappedBy' domain class property.
The existing documentation did not explain why you might want to use the
'mappedBy' property for many-to-one or one-to-one associations (only for
one-to-manies). This has now been added. I have also highlighted that a
developer can use values of "none" to indicate that an association is
unidirectional rather than bidirectional.