Updated the Simpson example to use 'ref'
In the current documentation, the bean definition surrounding the Simpson's characters doesn't run successfully. This is because bart and lisa are defined after marge, yet are referenced without 'ref'.

A quick scan didn't reveal any other violations of this rule, other than this example. An alternative solution would be to move lisa and bart above the marge bean definition.