Fix incorrect doc on `UrlConstraint` using `commons-validator`.
`UrlConstraint` uses its own `UrlValidator` and `DomainValidator`, and
does *not* use `commons-validator`.  This matters in the case you are
trying to get Grails to accept new TLDs, such as `.pizza` or `.wtf`:
upgrading `commons-validator` to the latest version won't help, since it
is not used to validate URLs, and the Grails `DomainValidator` maintains
its own TLD list.