Qualify use of media-range
Considered raising a Jira for this as I am not sure this is the best place for this note.  The "Versioning REST resources" section might benefit from further expansion or, can we just reference another persons' work, e.g. http://barelyenough.org/blog/2008/05/versioning-rest-web-services/.  I guess we do not want to full the grails manual with everything about web services.  I was worried that the examples are not representative of actual practice.