Update integration test documentation for transaction changes
As of v3.0.10, the transactional behavior of integration tests is
different than that presented in the documentation. Commit
cc2f2ed86d18fded831242d0c8d2d1ee885fc8a5 in grails-core introduced the
ability to persist objects in the setup() method of integration tests
via the default @Rollback annotation.

That commit essentially wrapped the contents of the setup() method in
its own transaction that is not rolled back.