Update run-app.adoc (#698)
In this proposal I've kept the term "Embedded containers" even though Spring Boot mostly seems to refer to "Embedded web servers". At least with regards to changing between servers. To adhere to Spring Boot documentation changing to the latter might be a good idea in this section, however with this minor change at least the user is sent to the relevant URL and not a 404 page.