fix(ci): Update snapshot publishing
Snapshots for the active branch were not published
because of a condition in the github workflow.

The SKIP_SNAPSHOT env var should suffice to
prevent snapshot publishing of older branches.

Also formatted the SKIP_SNAPSHOT conditions,
so it is easier to read and maintain, especially
when more branches are added.